refactor: delete MemoryGate and tighten load-path surfaces
`ModelLoadPlan` + `LoadDenyPolicy` (Stages 3-4) fully subsume the legacy
`MemoryGate` pre-load check. This commit removes the dead surface:

- Delete `MemoryGate.swift` and its two test files.
- Delete `ModelLoadPlanParityTests` (compared gates that no longer exist).
- Remove `InferenceService.memoryGate`, `ModelLifecycleCoordinator.memoryGate`,
  and the `ChatViewModel` auto-wiring block.
- Deprecate `loadModel(from:contextSize:)`; callers should build a
  `ModelLoadPlan` and use `loadModel(from:plan:)`.
- Add `LoadDenyPolicyInferenceServiceTests` covering the deny-policy
  invariants (throwError / warnOnly / custom) directly via
  `loadModel(from:plan:)`, without the retired gate.
- Update docs (DocC, README, RELIABILITY.md, GettingStarted) to describe
  the plan + deny-policy flow.

Keeps `DeviceCapabilityService.canLoadModel` + `safeContextSize` pending
migration in #423.
