--- Tests/BaseChatInferenceTests/ModelLoadPlanParityTests.swift
@@ -1,215 +1,0 @@
-import XCTest
-@testable import BaseChatInference
-
-/// Parity harness between `ModelLoadPlan` and the legacy gates (`DeviceCapabilityService.canLoadModel`
-/// + `MemoryGate.check`).
-///
-/// This is the safety rail for Stages 2–4 of the load-path refactor. The invariant it
-/// enforces is the **union** of the legacy checks: if *either* legacy gate rejects,
-/// the plan must deny; if *both* accept, the plan must not deny.
-///
-/// The two legacy gates use different denominators (physical vs available memory), so
-/// they sometimes disagree. Those one-side-rejects cases are the overlap region where
-/// the plan's new unified math supersedes — we record them but do not assert on them.
-///
-/// Grid sizing: we curate a ~60-case grid that exercises both the agreement region
-/// (clear pass / clear fail) and the rejection region, rather than the full 6 000
-/// cross-product. Selection criterion: sweep one dimension at a time around the
-/// agreement boundaries.
-final class ModelLoadPlanParityTests: XCTestCase {
-
-    private let oneGB: UInt64 = 1_073_741_824
-
-    /// A single test input row. Kept as a tuple rather than a struct to make the
-    /// curated grid readable inline.
-    private struct Row {
-        let fileSize: UInt64
-        let strategy: MemoryStrategy
-        let requested: Int
-        let trained: Int?
-        let kvPerTok: UInt64
-        let available: UInt64
-        let physical: UInt64
-    }
-
-    /// The curated grid. We take 5 cross-cut slices rather than a full cross-product:
-    ///   1. Small models, varied memory — should mostly agree and allow.
-    ///   2. Large models on small devices — should agree and deny.
-    ///   3. Boundary-of-resident-fit cases — tests the agreement region carefully.
-    ///   4. Large context on tight memory — exercises the plan's context-clamp path.
-    ///   5. A handful of the full-grid's disagreement cases — the overlap region.
-    private func buildGrid() -> [Row] {
-        var rows: [Row] = []
-
-        // Slice 1: Small 0.5 GB model — trivially fits on anything.
-        for avail in [oneGB, 2 * oneGB, 4 * oneGB, 8 * oneGB] {
-            for strategy in [MemoryStrategy.resident, .mappable] {
-                rows.append(Row(
-                    fileSize: oneGB / 2, strategy: strategy,
-                    requested: 2048, trained: 8192, kvPerTok: 8_192,
-                    available: avail, physical: 8 * oneGB
-                ))
-            }
-        }
-
-        // Slice 2: Large 16 GB model — doesn't fit on consumer devices.
-        for avail in [oneGB, 2 * oneGB, 4 * oneGB] {
-            for strategy in [MemoryStrategy.resident, .mappable] {
-                rows.append(Row(
-                    fileSize: 16 * oneGB, strategy: strategy,
-                    requested: 4096, trained: 8192, kvPerTok: 8_192,
-                    available: avail, physical: 8 * oneGB
-                ))
-            }
-        }
-
-        // Slice 3: 4 GB models around the boundary — both strategies against both RAM sizes.
-        for avail in [2 * oneGB, 4 * oneGB, 8 * oneGB] {
-            for strategy in [MemoryStrategy.resident, .mappable] {
-                for physical in [UInt64(8) * oneGB, UInt64(16) * oneGB] {
-                    rows.append(Row(
-                        fileSize: 4 * oneGB, strategy: strategy,
-                        requested: 4096, trained: 8192, kvPerTok: 8_192,
-                        available: avail, physical: physical
-                    ))
-                }
-            }
-        }
-
-        // Slice 4: Large-context requests. Plan clamps, legacy doesn't care about context
-        // for MemoryGate but canLoadModel ignores requested too — so these should align
-        // with slice 3 on verdict.
-        for requested in [32_000, 128_000] {
-            for kv in [UInt64(8_192), UInt64(131_072)] {
-                rows.append(Row(
-                    fileSize: 4 * oneGB, strategy: .mappable,
-                    requested: requested, trained: 131_072, kvPerTok: kv,
-                    available: 8 * oneGB, physical: 16 * oneGB
-                ))
-                rows.append(Row(
-                    fileSize: 4 * oneGB, strategy: .mappable,
-                    requested: requested, trained: 131_072, kvPerTok: kv,
-                    available: 2 * oneGB, physical: 8 * oneGB
-                ))
-            }
-        }
-
-        // Slice 5: Intentional disagreement region — available >> 0.70*physical
-        // (e.g., macOS reports lots of free memory on a lightly-loaded 8 GB Mac).
-        rows.append(Row(
-            fileSize: 4 * oneGB, strategy: .resident,
-            requested: 4096, trained: 8192, kvPerTok: 8_192,
-            available: 7 * oneGB, physical: 8 * oneGB
-        ))
-        // And the reverse — plenty of physical, but current available is low.
-        rows.append(Row(
-            fileSize: 4 * oneGB, strategy: .resident,
-            requested: 4096, trained: 8192, kvPerTok: 8_192,
-            available: oneGB / 2, physical: 32 * oneGB
-        ))
-
-        return rows
-    }
-
-    /// Collapses a `MemoryGate.Verdict` to "is this a rejection?" because the parity
-    /// invariant treats warn as non-rejection.
-    private func gateRejects(_ verdict: MemoryGate.Verdict) -> Bool {
-        if case .deny = verdict { return true }
-        return false
-    }
-
-    /// Evaluates the union invariant across the curated grid.
-    ///
-    /// The strongest form of the invariant (both legacy reject ⇒ plan denies, both
-    /// accept ⇒ plan doesn't deny) does **not** hold perfectly — the plan's new
-    /// unified math intentionally diverges in two places:
-    ///   1. Mappable strategy: plan caps resident at min(fileSize/4, 1 GB) whereas
-    ///      legacy MemoryGate uses uncapped fileSize/4. For very large mmap files
-    ///      the plan is *more* permissive than legacy.
-    ///   2. Boundary arithmetic: when residentBudget == available, the plan floors
-    ///      the context at 1 which causes total to micro-overshoot and deny, while
-    ///      both legacy gates warn/accept at that same boundary.
-    ///
-    /// Stages 2–4 will need to choose a reconciliation (either widen the plan's
-    /// mappable estimate, or relax the boundary verdict). Until then, we record
-    /// divergence counts rather than asserting on them, and assert only the *weak*
-    /// invariant: when the plan *and* both legacy gates all agree, nothing surprises.
-    func test_unionInvariant_recordsDivergenceWithoutFailing() {
-        let grid = buildGrid()
-        XCTAssertGreaterThanOrEqual(grid.count, 30, "grid too small to be meaningful")
-
-        var bothRejectPlanAllows: [Row] = []
-        var bothAcceptPlanDenies: [Row] = []
-        var oneRejectsOneAccepts = 0
-        var allAgree = 0
-
-        for row in grid {
-            let capacityService = DeviceCapabilityService(physicalMemory: row.physical)
-            let canLoad = capacityService.canLoadModel(estimatedMemoryBytes: row.fileSize)
-
-            let availableLocal = row.available
-            let gate = MemoryGate(
-                availableMemoryBytes: { availableLocal },
-                physicalMemoryBytes: row.physical,
-                denyBehavior: .warnOnly
-            )
-            let gateVerdict = gate.check(modelFileSize: row.fileSize, strategy: row.strategy)
-
-            let plan = ModelLoadPlan.compute(inputs: ModelLoadPlan.Inputs(
-                modelFileSize: row.fileSize,
-                memoryStrategy: row.strategy,
-                requestedContextSize: row.requested,
-                trainedContextLength: row.trained,
-                kvBytesPerToken: row.kvPerTok,
-                availableMemoryBytes: row.available,
-                physicalMemoryBytes: row.physical,
-                absoluteContextCeiling: 128_000,
-                headroomFraction: 0.40
-            ))
-
-            let legacyCapacityRejects = !canLoad
-            let legacyGateRejects = gateRejects(gateVerdict)
-
-            switch (legacyCapacityRejects, legacyGateRejects) {
-            case (true, true):
-                if plan.verdict != .deny {
-                    bothRejectPlanAllows.append(row)
-                } else {
-                    allAgree += 1
-                }
-            case (false, false):
-                if plan.verdict == .deny {
-                    bothAcceptPlanDenies.append(row)
-                } else {
-                    allAgree += 1
-                }
-            default:
-                oneRejectsOneAccepts += 1
-            }
-        }
-
-        // Sanity that the grid exercised enough of the space to be meaningful.
-        XCTAssertGreaterThan(allAgree, 0, "grid produced no agreement cases — check grid coverage")
-
-        // Document the known divergences so a reviewer in Stage 2–4 sees the count
-        // before changing anything. We do not fail on them — that's the whole point
-        // of the safety rail being a tracking harness rather than a gate.
-        print("""
-        [ModelLoadPlan parity summary]
-          total grid rows: \(grid.count)
-          all three agree: \(allAgree)
-          one-side-rejects (plan unconstrained): \(oneRejectsOneAccepts)
-          both-legacy-reject, plan allows: \(bothRejectPlanAllows.count)
-          both-legacy-accept, plan denies: \(bothAcceptPlanDenies.count)
-        """)
-
-        // Weak invariant: divergence should be small relative to the grid. If Stage 2–4
-        // code accidentally inverts the plan's verdict, the divergence count will
-        // balloon and this guard fires.
-        let divergence = bothRejectPlanAllows.count + bothAcceptPlanDenies.count
-        XCTAssertLessThan(
-            divergence, grid.count / 2,
-            "plan diverged from both legacy gates on more than half the grid — likely regression"
-        )
-    }
-}
