+++ docs/RELIABILITY.md
@@ -16,7 +16,7 @@ Once the first response bytes arrive, parsing runs outside the retry wrapper. Th
 
 ## Memory behavior
 
-BaseChatKit has two separate memory defenses. The first is preflight admission control through [`MemoryGate`](../Sources/BaseChatInference/Services/MemoryGate.swift), which `InferenceService` consults only if the app sets `inferenceService.memoryGate`. For `resident` backends it estimates roughly the full model file size; for `mappable` backends it estimates roughly 25% of the file size; for `external` backends it always allows the load. The default gate throws on deny on iOS and only warns on macOS.
+BaseChatKit has two separate memory defenses. The first is preflight admission control through [`ModelLoadPlan`](../Sources/BaseChatInference/Services/ModelLoadPlan.swift), which is built by callers (or automatically by the drop-in UI) before a load commits. For `resident` backends it estimates roughly the full model file size plus KV cache; for `mappable` backends it estimates roughly 25% of the file size plus KV; for `external` backends it always allows the load. When the plan returns a `.deny` verdict, `InferenceService` applies [`LoadDenyPolicy`](../Sources/BaseChatInference/Services/LoadDenyPolicy.swift) — `.throwError` (iOS default), `.warnOnly` (macOS default), or `.custom` with full access to the plan's `reasons`.
 
 The second is runtime memory-pressure handling. [`MemoryPressureHandler`](../Sources/BaseChatInference/Services/MemoryPressureHandler.swift) reports `.nominal`, `.warning`, and `.critical`, but it does not unload anything by itself. In the drop-in UI, [`ChatViewModel.handleMemoryPressure()`](../Sources/BaseChatUI/ViewModels/ChatViewModel+MemoryPressure.swift) shows a warning banner at `.warning` and stops generation plus unloads the model at `.critical`. If an app never starts monitoring or never forwards those events into the view model, there is no automatic unload path. Also, despite the handler comment mentioning "pause heavy work," the current UI behavior does not pause generation at `.warning`; only `.critical` triggers a stop and unload.
